Extract common codes to WebAppBaseWrapper
